Move Server enum
