+++ dadguide/common/enums.py
@@ -1,0 +1,5 @@
+from typing import Literal
+
+DEFAULT_SERVER = "COMBINED"
+SERVERS = ["NA", "COMBINED"]
+Server = Literal["COMBINED", "NA"]
