+++ dadguide/database_context.py
@@ -5,7 +5,8 @@ from .models.awoken_skill_model import AwokenSkillModel
 from .models.dungeon_model import DungeonModel
 from .models.scheduled_event_model import ScheduledEventModel
 from .models.series_model import SeriesModel
-from .monster_graph import MonsterGraph, Server, DEFAULT_SERVER
+from .monster_graph import MonsterGraph
+from .common.enums import DEFAULT_SERVER, Server
 
 SCHEDULED_EVENT_QUERY = """SELECT
   schedule.*,
