+++ dadguide/idtest_mixin.py
@@ -6,7 +6,7 @@ import tsutils
 from redbot.core import commands, Config, checks
 from redbot.core.utils.chat_formatting import box, pagify
 
-from dadguide.monster_graph import Server, DEFAULT_SERVER
+from dadguide.common.enums import DEFAULT_SERVER, Server
 from dadguide.monster_index import MonsterIndex
