+++ dadguide/monster_index.py
@@ -9,7 +9,8 @@ import tsutils
 from redbot.core.utils import AsyncIter
 
 from .models.monster_model import MonsterModel
-from .monster_graph import MonsterGraph, Server, DEFAULT_SERVER
+from .monster_graph import MonsterGraph
+from .common.enums import DEFAULT_SERVER, Server
 from .token_mappings import *
 
 SHEETS_PATTERN = 'https://docs.google.com/spreadsheets/d/1EoZJ3w5xsXZ67kmarLE4vfrZSIIIAfj04HXeZVST3eY' \
